+++ src/app/admin-module/components/source-detail/source-detail.component.ts
@@ -2,12 +2,11 @@ import { Component, OnDestroy, OnInit, ViewEncapsulation } from '@angular/core';
 import { NonNullableFormBuilder, Validators } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
 import { Subject, takeUntil } from 'rxjs';
-
-import { DataCollectionSource } from '../models/data-collection-source.interface';
-import { DataCollectionService } from '../services/data-collection.service';
-import { LanguageService } from '../../shared/services/language-service/language.service';
-import locale from '../admin.locale.json';
-import sharedLocale from '../../shared/general.locale.json';
+import { DataCollectionSource } from '../../models/data-collection-source.interface';
+import { DataCollectionService } from '../../services/data-collection.service';
+import { LanguageService } from '../../../shared/services/language-service/language.service';
+import locale from '../../admin.locale.json';
+import sharedLocale from '../../../shared/general.locale.json';
 
 @Component({
   selector: 'app-source-detail',