+++ src/app/admin-module/functions/parse-null.function.ts
@@ -2,7 +2,7 @@
  * If value is 'null' or 'undefined' then return 'null', otherwise return string value of input value
  * @param value
  */
-export function parseNull(value: unknown) {
+export function parseNull(value: any) {
   if (value === null) {
     return null;
   } else if (value === undefined) {