+++ src/app/component-modules/autocomplete/components/autocomplete.component.ts
@@ -1,7 +1,6 @@
 import { Component, Input, OnDestroy, OnInit, Output, ViewEncapsulation } from '@angular/core';
 import { NonNullableFormBuilder } from '@angular/forms';
 import { BehaviorSubject, Subject, takeUntil } from 'rxjs';
-
 import { LanguageService } from '../../../shared/services/language-service/language.service';
 
 @Component({
@@ -10,7 +9,8 @@ import { LanguageService } from '../../../shared/services/language-service/langu
     <div class="block relative w-full">
       <div class="flex rounded-md">
         <input
-          class="block flex-1 w-full p-4 bg-transparent text-gray-900 h-12 rounded border outline-none"
+          class="block flex-1 w-full p-4 bg-transparent text-gray-900 h-12 outline-none"
+          [class]="inputClassName"
           [formControl]="inputValue"
           [placeholder]="placeholder"
           [matAutocomplete]="ingredientList"
@@ -43,6 +43,7 @@ export class AutocompleteComponent implements OnInit, OnDestroy {
   @Input() searchProp: string | number | undefined;
   /** Placeholder text */
   @Input() placeholder = '';
+  @Input() inputClassName = '';
   /** On autocomplete suggestion item select */
   @Output() onSelect = this.selected.asObservable();
 
@@ -81,14 +82,14 @@ export class AutocompleteComponent implements OnInit, OnDestroy {
   }
 
   /**
-   * Select item from suggestion
-   * @param item Selected item
+   * Selected option from Autocomplete
+   * @param selectedOption
    */
-  protected select(item: any) {
-    if (!item) {
+  protected select(selectedOption: any) {
+    if (!selectedOption) {
       return;
     }
-    this.selected.next(item);
+    this.selected.next(selectedOption);
     this.inputValue.reset();
   }
 }