+++ src/app/search-module/components/recipe/recipe.component.ts
@@ -8,23 +8,11 @@ import {
 } from '@angular/core';
 import { PageEvent } from '@angular/material/paginator';
 import { catchError, of, Subject, takeUntil, tap } from 'rxjs';
-import { LanguageObject } from '../../../language-switch-module/models/language-object.type';
-import generalLocale from '../../../shared/general.locale.json';
 import { ResponseList } from '../../../shared/models/response-list.type';
 import { LanguageService } from '../../../shared/services/language-service/language.service';
-
 import { Recipe } from '../../models/recipe.interface';
 import { RecipeService } from '../../services/recipe.service';
 import locale from './recipe.locale.json';
-
-const LOCALE_TEXT: LanguageObject = {
-  cs: {
-    MINUTES: 'minut',
-  },
-  en: {
-    MINUTES: 'minutes',
-  },
-};
 
 @Component({
   selector: 'app-recipe',
@@ -57,7 +45,7 @@ const LOCALE_TEXT: LanguageObject = {
               >
                 <mat-icon fontIcon="schedule" class="mr-1"></mat-icon>
                 <span class="text-sm text-gray-500 font-medium">
-                  {{ recipe.preparation_time + ' ' + LOCALE_TEXT[lang.language]['MINUTES'] }}
+                  {{ recipe.preparation_time + ' ' + locale[lang.language].MINUTES }}
                 </span>
               </div>
               <!--calories-->
@@ -130,11 +118,9 @@ const LOCALE_TEXT: LanguageObject = {
 export class RecipeComponent implements OnInit, OnDestroy {
   private readonly subs = new Subject<boolean>();
   protected readonly locale = locale;
-  protected readonly generalLocale = generalLocale;
   protected recipeList: ResponseList<Recipe> | undefined;
   protected pageIndex = 0;
   protected readonly pageSize = 5;
-  protected readonly LOCALE_TEXT = LOCALE_TEXT;
 
   constructor(
     protected readonly lang: LanguageService,