+++ src/app/search-module/components/search-page/search-page.component.ts
@@ -1,6 +1,5 @@
 import { Component, ViewEncapsulation } from '@angular/core';
 import { Title } from '@angular/platform-browser';
-
 import { APP_PAGE_TITLE } from "../../../app.settings";
 import { LanguageService } from "../../../shared/services/language-service/language.service";
 import locale from './search.locale.json';