+++ src/app/search-module/models/ingredient-category.interface.ts
@@ -2,6 +2,7 @@ import { Locale } from './locale.interface';
 import { Ingredient } from './ingredient.interface';
 
 export interface IngredientCategory {
+  id: number;
   name: string;
   locale: Locale;
   ingredients: Ingredient[];