+++ src/app/search-module/services/ingredient-dialog.service.ts
@@ -1,11 +1,36 @@
 import { Injectable } from '@angular/core';
 import { Ingredient } from '../models/ingredient.interface';
 import { IngredientCategory } from '../models/ingredient-category.interface';
+import { CategoryService } from './category.service';
 
 @Injectable()
 export class IngredientDialogService {
+  private categoryList: IngredientCategory[] = [];
   /** List of selected Ingredient */
   private selectedIngredientList: Ingredient[] = [];
+
+  constructor(private readonly categoryService: CategoryService) {
+    // Get ingredient categories - memory leak?
+    this.categoryService.loadCategories().subscribe({
+      next: (data) => {
+        this.categoryList = data;
+      },
+    });
+  }
+
+  /**
+   * Function to find the category containing the ingredient
+   * @param ingredient
+   */
+  private findCategory(ingredient: Ingredient) {
+    const categoryID = ingredient.category_id;
+    return this.categoryList.find((cat) => cat.id === categoryID);
+  }
+
+  private findIngredient(selectedIngredient: Ingredient, category: IngredientCategory) {
+    const ingredientList = category.ingredients;
+    return ingredientList.find((ing) => ing.id === selectedIngredient.id);
+  }
 
   /** List of selected Ingredients */
   get selectedList() {
@@ -21,17 +46,11 @@ export class IngredientDialogService {
    * Switch state of selected Ingredient
    * @param selectedIngredient
    * @param nextState
-   * @param categoryList
    */
-  toggleIngredientSelection(
-    selectedIngredient: Ingredient,
-    nextState: boolean,
-    categoryList: IngredientCategory[]
-  ) {
+  toggleIngredientSelection(selectedIngredient: Ingredient, nextState: boolean) {
     selectedIngredient.selected = nextState;
-
     // re-create new selection list above search-bar
-    this.selectedIngredientList = categoryList.reduce((result: Ingredient[], category) => {
+    this.selectedIngredientList = this.categoryList.reduce((result: Ingredient[], category) => {
       const selected = category.ingredients.filter((ingredient) => ingredient.selected);
       return result.concat(selected); // or [...selectedList, ...selected] for an alternative approach
     }, []);
@@ -40,18 +59,16 @@ export class IngredientDialogService {
   /**
    * Search through Ingredient Categories to find selected Ingredient item.
    * @param selectedIngredient
-   * @param categoryList
    */
-  ingredientSelect(selectedIngredient: Ingredient, categoryList: IngredientCategory[]) {
-    // TODO i dont like the way of nested documents, there must be a way for better solution
-    for (const category of categoryList) {
-      for (const ingredient of category.ingredients) {
-        if (ingredient.id === selectedIngredient.id) {
-          // change state to single item
-          this.toggleIngredientSelection(ingredient, true, categoryList);
-          break; // Exit the function after changing state
-        }
-      }
+  ingredientSelect(selectedIngredient: Ingredient) {
+    const foundCategory = this.findCategory(selectedIngredient);
+    if (!foundCategory) {
+      return;
     }
+    const foundIngredient = this.findIngredient(selectedIngredient, foundCategory);
+    if (!foundIngredient) {
+      return;
+    }
+    this.toggleIngredientSelection(foundIngredient, true);
   }
 }